requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pydantic
python-multipart
openai
email-validator
gpt-researcher
requests>=2.25.1 